fix : 댓글 채택시 질문 게시물은 삭제 불가하게 수정

+++ src/main/java/CamNecT/CamNecT_Server/domain/community/service/PostServiceImpl.java
@@ -117,6 +117,11 @@ public class PostServiceImpl implements PostService {
             throw new IllegalArgumentException("forbidden");
         }
 
+        if (post.getBoard().getCode() == BoardCode.QUESTION
+                && acceptedCommentsRepository.existsByPost_Id(postId)) {
+            throw new IllegalArgumentException("cannot delete accepted question");
+        }
+
         post.deleteSoft();
 
         postAttachmentsRepository.softDeleteByPostId(postId);
